Update SDK tests
* Add some test for MVC 3.1 \ Blazor 3.1
* Remove tests and testapps for 1.1 \ 2.2
