--- src/Razor/Microsoft.NET.Sdk.Razor/test/IntegrationTests/BuildIntegrationTest22.cs
@@ -1,16 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-namespace Microsoft.AspNetCore.Razor.Design.IntegrationTests
-{
-    public class BuildIntegrationTest22 : BuildIntegrationTestLegacy
-    {
-        public BuildIntegrationTest22(LegacyBuildServerTestFixture buildServer)
-            : base(buildServer)
-        {
-        }
-
-        public override string TestProjectName => "SimpleMvc22";
-        public override string TargetFramework => "netcoreapp2.2";
-    }
-}
