+++ src/Razor/Microsoft.NET.Sdk.Razor/test/IntegrationTests/MvcBuildIntegrationTest31.cs
@@ -1,0 +1,16 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace Microsoft.AspNetCore.Razor.Design.IntegrationTests
+{
+    public class MvcBuildIntegrationTest31 : MvcBuildIntegrationTestLegacy
+    {
+        public MvcBuildIntegrationTest31(LegacyBuildServerTestFixture buildServer)
+            : base(buildServer)
+        {
+        }
+
+        public override string TestProjectName => "SimpleMvc31";
+        public override string TargetFramework => "netcoreapp3.1";
+    }
+}
