+++ src/Razor/test/testassets/Directory.Build.props
@@ -13,6 +13,9 @@
     <EnableCustomReferenceResolution>false</EnableCustomReferenceResolution>
 
     <DefaultNetCoreTargetFramework>net5.0</DefaultNetCoreTargetFramework>
+
+    <RepoRoot Condition="'$(BuildingTestAppsIndependently)' == 'true' AND '$(RepoRoot)' == ''">$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory)..\, 'global.json'))\</RepoRoot>
+    <RazorSdkArtifactsDirectory>$(RepoRoot)artifacts\bin\Microsoft.NET.Sdk.Razor\</RazorSdkArtifactsDirectory>
   </PropertyGroup>
 
   <Import Project="$(RepoRoot)eng\Versions.props" />
