+++ src/Razor/test/testassets/RestoreTestProjects/RestoreTestProjects.csproj
@@ -4,12 +4,10 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <!-- Intentionally not including SimpleMvc11 and SimpleMvc11NetFx here because they are restored in a separate target. -->
     <ProjectReference Include="..\SimpleMvc21\SimpleMvc21.csproj" />
-    <ProjectReference Include="..\SimpleMvc22\SimpleMvc22.csproj" />
-    <ProjectReference Include="..\SimpleMvc22NetFx\SimpleMvc22NetFx.csproj" />
+    <ProjectReference Include="..\SimpleMvc31\SimpleMvc31.csproj" />
+    <ProjectReference Include="..\blazor31\blazor31.csproj" />
     <ProjectReference Include="..\ClassLibraryMvc21\ClassLibraryMvc21.csproj" />
-
     <ProjectReference Include="..\AppWithP2PReference\AppWithP2PReference.csproj" />
     <ProjectReference Include="..\AppWithPackageAndP2PReferenceAndRID\AppWithPackageAndP2PReferenceAndRID.csproj" />
     <ProjectReference Include="..\ClassLibrary\ClassLibrary.csproj" />
@@ -24,8 +22,8 @@
     <ProjectReference Include="..\SimplePages\SimplePages.csproj" />
     <ProjectReference Include="..\AppWithPackageAndP2PReference\AppWithPackageAndP2PReference.csproj" />
 
-    <ProjectReference Include="..\..\Microsoft.AspNetCore.Razor.Test.MvcShim.ClassLib\Microsoft.AspNetCore.Razor.Test.MvcShim.ClassLib.csproj"/>
-    <ProjectReference Include="..\..\Microsoft.AspNetCore.Razor.Test.ComponentShim\Microsoft.AspNetCore.Razor.Test.ComponentShim.csproj"/>
+    <ProjectReference Include="..\..\Microsoft.AspNetCore.Razor.Test.MvcShim.ClassLib\Microsoft.AspNetCore.Razor.Test.MvcShim.ClassLib.csproj" />
+    <ProjectReference Include="..\..\Microsoft.AspNetCore.Razor.Test.ComponentShim\Microsoft.AspNetCore.Razor.Test.ComponentShim.csproj" />
   </ItemGroup>
 
 </Project>
