--- src/Razor/test/testassets/SimpleMvc11/Views/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-﻿@using SimpleMvc11
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@inject DateTime TheTime
