--- src/Razor/test/testassets/SimpleMvc11/Views/_ViewStart.cshtml
@@ -1,3 +1,0 @@
-﻿@{
-    Layout = "_Layout";
-}
