--- src/Razor/test/testassets/SimpleMvc11NetFx/SimpleMvc11NetFx.csproj
@@ -1,19 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <!--
-    This project references a shipped version of MVC and should not reference local builds of
-    the CodeGeneration targets, rzc, or any of the test shims.
-  -->
-
-
-  <PropertyGroup>
-    <TargetFramework>net461</TargetFramework>
-  </PropertyGroup>
-
-  <!-- Test Placeholder -->
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore" Version="1.1.7" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="1.1.8" />
-  </ItemGroup>
-</Project>
