--- src/Razor/test/testassets/SimpleMvc11NetFx/Views/_ViewImports.cshtml
@@ -1,2 +1,0 @@
-﻿@using SimpleMvc11NetFx
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
