--- src/Razor/test/testassets/SimpleMvc22/SimpleMvc22.csproj
@@ -1,23 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <!--
-    This project references a shipped version of MVC. It will reference the local build of Tasks but not the compiler or any test shims.
-  -->
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-  </PropertyGroup>
-
-  <!-- Test Placeholder -->
-
-  <PropertyGroup Condition="'$(RunningAsTest)' == ''">
-    <!-- We don't want to run build server when not running as tests. -->
-    <UseRazorBuildServer>false</UseRazorBuildServer>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
-  </ItemGroup>
-
-</Project>
