--- src/Razor/test/testassets/SimpleMvc22/Views/_ViewImports.cshtml
@@ -1,5 +1,0 @@
-@using SimpleMvc
-@using SimpleMvc.Models
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@namespace SimpleMvc21
-@inject DateTime TheTime
