--- src/Razor/test/testassets/SimpleMvc22NetFx/Models/ErrorViewModel.cs
@@ -1,11 +1,0 @@
-using System;
-
-namespace SimpleMvc.Models
-{
-    public class ErrorViewModel
-    {
-        public string RequestId { get; set; }
-
-        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-    }
-}
