--- src/Razor/test/testassets/SimpleMvc22NetFx/SimpleTagHelper.cs
@@ -1,8 +1,0 @@
-using Microsoft.AspNetCore.Razor.TagHelpers;
-
-namespace SimpleMvc
-{
-    public class SimpleTagHelper : TagHelper
-    {
-    }
-}
