--- src/Razor/test/testassets/SimpleMvc22NetFx/Views/_ViewImports.cshtml
@@ -1,4 +1,0 @@
-@using SimpleMvc
-@using SimpleMvc.Models
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@namespace SimpleMvc21
