--- src/Razor/test/testassets/SimpleMvc22NetFx/Views/_ViewStart.cshtml
@@ -1,3 +1,0 @@
-﻿@{
-    Layout = "_Layout";
-}
