+++ src/Razor/test/testassets/SimpleMvc31/Models/ErrorViewModel.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace SimpleMvc31.Models
+{
+    public class ErrorViewModel
+    {
+        public string RequestId { get; set; }
+
+        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+    }
+}
