+++ src/Razor/test/testassets/SimpleMvc31/Program.cs
@@ -1,0 +1,13 @@
+
+namespace SimpleMvc31
+{
+    public class Program
+    {
+        public static void Main(string[] args)
+        {
+            // Just make sure we have a reference to the MvcShim
+            var t = typeof(Microsoft.AspNetCore.Mvc.IActionResult);
+            System.Console.WriteLine(t.FullName);
+        }
+    }
+}
