+++ src/Razor/test/testassets/SimpleMvc31/Views/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+﻿@using SimpleMvc31
+@using SimpleMvc31.Models
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
