+++ src/Razor/test/testassets/SimpleMvc31/Views/_ViewStart.cshtml
@@ -1,0 +1,3 @@
+@{
+    Layout = "_Layout";
+}
