+++ src/Razor/test/testassets/blazor31/Pages/Index.razor
@@ -1,0 +1,6 @@
+﻿@page "/"
+
+<h1>Hello, world!</h1>
+
+Welcome to your new app.
+
