+++ src/Razor/test/testassets/blazor31/Shared/MainLayout.razor
@@ -1,0 +1,14 @@
+﻿@inherits LayoutComponentBase
+
+<div class="sidebar">
+</div>
+
+<div class="main">
+    <div class="top-row px-4">
+        <a href="https://docs.microsoft.com/aspnet/" target="_blank">About</a>
+    </div>
+
+    <div class="content px-4">
+        @Body
+    </div>
+</div>
